refactor(inquiry): rename views => Views
