+++ src/v2/Components/Inquiry/Views/InquiryArtistsInCollection.tsx
