+++ src/v2/Components/Inquiry/Views/InquiryBasicInfo.tsx
