+++ src/v2/Components/Inquiry/Views/InquiryConfirmation.tsx
