+++ src/v2/Components/Inquiry/__tests__/Views/InquiryCommercialInterest.jest.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { mount } from "enzyme"
-import { InquiryCommercialInterest } from "../../views/InquiryCommercialInterest"
+import { InquiryCommercialInterest } from "../../Views/InquiryCommercialInterest"
 import { useUpdateMyUserProfile } from "../../Hooks/useUpdateMyUserProfile"
 import { useInquiryContext } from "../../InquiryContext"
 import { flushPromiseQueue } from "v2/DevTools"
